moving canjs bits around
