+++ public/canjs/controls/select-report-type/select-report-type.js
@@ -1,6 +1,6 @@
-import { StacheElement, type, ObservableObject, ObservableArray, value } from "../can.js";
+import { StacheElement, type, ObservableObject, ObservableArray, value } from "../../../can.js";
 
-import {saveJSONToUrl,updateUrlParam} from "../shared/state-storage.js";
+import {saveJSONToUrl,updateUrlParam} from "../../routing/state-storage.js";
 
 import "../status-filter.js";
 
@@ -14,7 +14,7 @@ const booleanParsing = {
 
 const selectStyle = "bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 p-2 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
 
-import SimpleTooltip from "../shared/simple-tooltip.js";
+import SimpleTooltip from "../../ui/simple-tooltip/simple-tooltip";
 const TOOLTIP = new SimpleTooltip();
 document.body.append(TOOLTIP);
 
@@ -48,7 +48,7 @@ class ReportSelectionDropdown extends StacheElement {
 
 customElements.define("report-selection-dropdown", ReportSelectionDropdown);
 
-import { DROPDOWN_LABEL } from "../shared/style-strings.js";
+import { DROPDOWN_LABEL } from "../../../shared/style-strings.js";
 
 export class SelectReportType extends StacheElement {
     static view = `
