+++ public/canjs/controls/status-filter.js
@@ -1,6 +1,5 @@
-import { StacheElement, type, ObservableObject, ObservableArray } from "./can.js";
-import {updateUrlParam} from "./shared/state-storage.js";
-import "./shared/autocomplete/autocomplete.js";
+import { StacheElement, type, ObservableObject, ObservableArray } from "../../can.js";
+import "../ui/autocomplete/autocomplete.js";
 
 export class StatusFilter extends StacheElement {
     static view = `
