+++ public/canjs/controls/timeline-configuration/timing-calculation/timing-calculation.js
@@ -1,10 +1,8 @@
-import { StacheElement, type, ObservableObject, ObservableArray, value } from "../../can.js";
+import { StacheElement, type, ObservableObject, ObservableArray, value } from "../../../../can.js";
 
-import {updateUrlParam} from "../../shared/state-storage.js";
+import {updateUrlParam} from "../../../routing/state-storage.js";
 
-import { bitoviTrainingIssueData } from "../../examples/bitovi-training.js";
-
-import { getSimplifiedIssueHierarchy } from "../../stateful-data/jira-data-requests.js";
+import { getSimplifiedIssueHierarchy } from "../../../../stateful-data/jira-data-requests.js";
 
 const selectStyle = "bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 p-2 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
