+++ public/canjs/reports/group-grid/group-grid.js
@@ -1,5 +1,5 @@
 
-import { StacheElement, type, ObservableObject, stache } from "../../can.js";
+import { StacheElement, type, ObservableObject, stache } from "../../../can.js";
 
 const DAY = 1000*60*60*24;
 export class GroupGrid extends StacheElement {
