+++ public/canjs/ui/simple-tooltip/simple-tooltip.js
